requirements.txt
flask-sqlalchemy
flask-login
werkzeug
Flask==1.1.2
gunicorn==20.1.0
google-cloud-storage
schedule
facebook-business==10.0.0
facebook-sdk==3.1.0
facebookads==2.11.4
json5==0.9.5
requests
numpy
pandas

